Additional refactoring of correlation classes
