+++ src/frontend/matchers/ObservablesMatcher.cpp
@@ -28,7 +28,7 @@
 #include "core/observables/correlation/LayerwiseRadialEnumerator.h"
 #include "core/observables/correlation/PairAveragedCorrelation.h"
 #include "core/observables/DensityHistogram.h"
-#include "core/observables/ProbabilityEvolution.h"
+#include "core/observables/correlation/ProbabilityEvolution.h"
 
 #include "core/observables/correlation_functions/S110Correlation.h"
 #include "core/observables/correlation_functions/AxesAngle.h"
